Expanded the amount of code included in the transform example
The amount of code included from "//transformExample start" to "//transformExample end" was previously insufficient to demonstrate how the prefixes should be structured or the collections created. It resulted in code being included in the Geometry documentation which did not build. I have modified the code to use an immutable FeatureCollection view, and have expanded the amount of code used in the example. The Geometry documentation will now have code that builds.