GEOT-5191: Use GeographicLib-Java for geodesic calculations.
The affected classes are GeodeticCalculator and DefaultEllipsoid.  Some
of the tests were wrong and these needed to be fixed.  Now you can
calculate the distance between Wellington, 41.32S 174.81E, and
Salamanca, 40.96N 5.50W.

The latitude must lie in [-90,90] as before.  But now, any finite
longitude, azimuth, and distance is accepted by the GeodeticCalculator.
The calculated longitude and azimuth is in the range [-180,180), except
that the longitudes returned by getGeodeticPath are "unrolled" and so
vary continuously (this is governed by the GeodesicMask.LONG_UNROLL
flag).