Move `PartiallyOrderedSet`
To reuse the existing `PartiallyOrderedSet` in `FactoryRegistry`, it
must be moved into gt-metadata or one of its dependencies, maybe
adding a new one. No good candidate was found and since gt-xml
(where the set was originally used) transitively depends on
gt-metadata, moving it to the latter is a good temporary solution.