Remove synchronization on AttributeExpressionImpl property accessor keeping semantics
AttributeExpressionImpl synchronizes its lastAccessor's
getter and setter, and is noticeable under high concurrency.

Adding the volatile modified to the instance variable maintains
the semantics, synchronizing on the variable reference instead
of the container object's monitor, in a lock free way.

This improves AttributeExpressionImpl.evaluate()
from being 94% of the time blocked on getLastPropertyAccessor(),
to being blocked 23% of the time on the ArrayList.iterator()
java.lang.Reference$Lock GC aid when calling findPropertyAccessors().